requirements.txt
Flask==3.0.2
Flask-CORS==4.0.0
python-dotenv==1.0.1
PyMuPDF
cohere==4.45.0
pinecone-client==3.1.0
groq==0.4.2
werkzeug==3.0.1
gunicorn  